packages/python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "epaper-dithering"
version = "0.6.2"
description = "Dithering algorithms for e-paper/e-ink displays"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    {name = "g4bri3lDev", email = "admin@g4bri3l.de"}
]

keywords = ["epaper", "eink", "dithering", "image-processing", "display"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Graphics",
    "Typing :: Typed",
]

dependencies = [
    "pillow>=10.0.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
test = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0"
]
dev = [
    "mypy>=1.19.1",
    "ruff>=0.14.10"
]

[project.urls]
Homepage = "https://opendisplay.org"
Repository = "https://github.com/OpenDisplay-org/epaper-dithering"
Documentation = "https://github.com/OpenDisplay-org/epaper-dithering#readme"

[tool.hatch.build.targets.wheel]
packages = ["src/epaper_dithering"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pylint.format]
max-line-length = 120

[tool.pylint.basic]
# Allow single-letter and short uppercase names common in color science (L, C, Y, dL, dC, dH)
good-names = ["i", "j", "k", "x", "y", "z", "L", "C", "Y"]
variable-rgx = "^[a-zA-Z_][a-zA-Z0-9_]*$"
argument-rgx = "^[a-zA-Z_][a-zA-Z0-9_]*$"

[tool.pylint.design]
max-args = 8
max-positional-arguments = 8
max-locals = 40
max-returns = 10
max-branches = 15

[tool.pylint.messages_control]
disable = [
    "fixme",                    # TODO comments are fine
    "duplicate-code",           # algorithmic variants share structure by design
    "consider-using-enumerate", # numpy indexing patterns don't benefit from enumerate
    "consider-using-max-builtin", # explicit comparison is clearer for float edge cases
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "prek>=0.3.4",
    "pylint>=4.0.5",
]
